// src/api/generated/models/site-code-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Enterprise API
 * React + TypeScript + Spring Boot OpenAPI
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface SiteCodeDTO {
    'uidx'?: number;
    'centerId'?: number;
    'tenantId'?: number;
    'codeKind'?: string;
    'codeName'?: string;
    'label'?: string;
    'codeNumber'?: number;
    'codeValue'?: string;
    'value'?: string;
    'codeDesc'?: string;
    'codeSeq'?: number;
    'codeLang'?: string;
    'useFlag'?: number;
    'updateEmployeeId'?: string;
    'updateTime'?: string;
}